requirements.txt
google-generativeai
pandas
python-dotenv
streamlit